requirements.txt
python-dotenv
psycopg2-binary
yt-dlp
openai-whisper